# backend/docs/swagger.yaml
basePath: /
definitions:
...
  router.dhcpExclusionResponse:
    properties:
      endIp:
        type: string
      externalId:
        type: string
      id:
        type: string
      lastSyncedAt:
        type: string
      scopeId:
        type: string
      startIp:
        type: string
    type: object
  router.dhcpReservationResponse:
    properties:
      description:
        type: string
      externalId:
        type: string
      id:
        type: string
      ip:
        type: string
      lastSyncedAt:
        type: string
      mac:
        type: string
      name:
        type: string
      scopeId:
        type: string
    type: object
  router.dhcpScopeResponse:
    properties:
      description:
        type: string
      endRange:
        type: string
      externalId:
        type: string
      id:
        type: string
      lastError:
        type: string
      lastSyncedAt:
        type: string
      leaseDurationHours:
        type: integer
      leaseDurationSeconds:
        type: integer
      name:
        type: string
      serverId:
        type: string
      startRange:
        type: string
      state:
        type: string
      subnet:
        type: string
      syncStatus:
        type: string
    type: object
...
  router.errorDocResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  router.serverResponse:
    properties:
      agentUrl:
        type: string
      apiKey:
        type: string
      failureCount:
        type: integer
      host:
        type: string
      id:
        type: string
      lastChecked:
        type: string
      name:
        type: string
      role:
        type: string
      status:
        type: string
      tlsInsecure:
        type: boolean
    type: object
...
  router.statusResponse:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: ZoneLease Windows DNS / DHCP 统一管理控制台后端 API，提供认证、服务器登记、DNS/DHCP 资源管理、刷新事件和审计查询接口。
  title: ZoneLease API
  version: "1.0"
paths:
...
  /api/dhcp/exclusions:
    post:
      consumes:
      - application/json
      description: 转发到 DHCP Agent 创建 Windows DHCP 排除范围，成功后按当前作用域延迟合并创建 runtime.refresh.dhcp.scope
        局部刷新任务。
      parameters:
      - description: DHCP 排除范围参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.dhcpExclusionResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.dhcpExclusionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 创建 DHCP 排除范围
      tags:
      - DHCP
...
  /api/dhcp/reservations:
    post:
      consumes:
      - application/json
      description: 转发到 DHCP Agent 创建 Windows DHCP 保留地址，成功后按当前作用域延迟合并创建 runtime.refresh.dhcp.scope
        局部刷新任务。
      parameters:
      - description: DHCP 保留地址参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.dhcpReservationResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.dhcpReservationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 创建 DHCP 保留地址
      tags:
      - DHCP
...
  /api/dhcp/scopes:
    post:
      consumes:
      - application/json
      description: 转发到 DHCP Agent 创建 Windows DHCP 作用域，成功后按当前作用域延迟合并创建 runtime.refresh.dhcp.scope
        局部刷新任务。
      parameters:
      - description: DHCP 作用域参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.dhcpScopeResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.dhcpScopeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 创建 DHCP 作用域
      tags:
      - DHCP
...
  /api/servers:
    post:
      consumes:
      - application/json
      description: 保存前会检查 Agent 名称、接口地址是否重复和角色是否合法；Agent 连通性由前端先调用 /api/servers/probe
        完成，保存接口不再重复访问 Agent。
      parameters:
      - description: 服务器参数，包含 name、role、agentUrl、可选 apiKey 和可选 tlsInsecure
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/router.serverResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/router.serverResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 添加服务器
      tags:
      - Servers
  /api/servers/{id}:
    delete:
      parameters:
      - description: 服务器 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.statusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/router.errorDocResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.errorDocResponse'
      security:
      - BearerAuth: []
      summary: 删除服务器
      tags:
      - Servers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"